Separate some auth related code to features/auth
Change-type: patch
